requirements.txt
appdirs==1.4.3
arabic-reshaper==2.1.1
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
CacheControl==0.12.6
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cssselect==1.1.0
cssselect2==0.4.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.6
django-braces==1.14.0
django-ckeditor==6.0.0
django-contrib-comments==2.0.0
django-etc==1.3.0
django-filter==2.4.0
django-hitcount==1.3.3
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.11.0
django-robots==4.0
django-star-ratings==0.9.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
django-xhtml2pdf==0.0.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.16
drf-pdf==0.2.0
future==0.18.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.7.0
jdatetime==3.6.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.2
Markdown==3.3.3
markdown2==2.4.0
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.1.0
progress==1.5
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pymemcache==3.4.1
pyparsing==2.4.6
PyPDF2==1.26.0
Pyphen==0.10.0
pyquery==1.4.3
python-bidi==0.4.2
python-decouple==3.4
python-monkey-business==1.0.0
pytoml==0.1.21
pytz==2021.1
readtime==1.1.1
reportlab==3.5.59
requests==2.22.0
retrying==1.3.3
six==1.14.0
sorl-thumbnail==12.7.0
soupsieve==2.1
sqlparse==0.4.1
swapper==1.1.2.Movie1
tinycss2==1.1.0
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5